Reorganize WDFN code into its own dir
