+++ assets/rollup.config.js
@@ -55,7 +55,8 @@ const getBundleConfig = function (src, dest) {
         treeshake: env === 'production'
     };
 
-    if (src == 'src/scripts/networks/index.js'){
+    if (src == 'src/scripts/networks/index.js' ||
+        src == 'src/scripts/wdfn-home/index.js'){
         configMap['external'] = {
            window: 'window'
         };
@@ -66,7 +67,8 @@ const getBundleConfig = function (src, dest) {
 
 module.exports = [
     getBundleConfig('src/scripts/index.js', 'dist/bundle.js'),
-    getBundleConfig('src/scripts/networks/index.js', 'dist/network-bundle.js')
+    getBundleConfig('src/scripts/networks/index.js', 'dist/network-bundle.js'),
+    getBundleConfig('src/scripts/wdfn-home/index.js', 'dist/wdfn-bundle.js')
 ];
