+++ assets/src/scripts/index.js
@@ -14,16 +14,12 @@ import {attachToNode as EmbedComponent} from './components/embed';
 import {attachToNode as DailyValueHydrographComponent} from './components/daily-value-hydrograph';
 import {attachToNode as HydrographComponent} from './components/hydrograph';
 import {attachToNode as MapComponent} from './components/map';
-import {attachToNode as USMapComponent} from './components/usmap';
-import {attachToNode as WDFNParamFilters} from './components/wdfn-param-filters';
 
 const COMPONENTS = {
     embed: EmbedComponent,
     'dv-hydrograph': DailyValueHydrographComponent,
     hydrograph: HydrographComponent,
-    map: MapComponent,
-    'us-map': USMapComponent,
-    'wdfn-param-filters': WDFNParamFilters
+    map: MapComponent
 };
 
 const load = function () {
