--- assets/src/scripts/selectors/waterquality-selector.js
@@ -1,4 +1,0 @@
-import { createSelector } from 'reselect';
-
-export const Sites = state => state.waterquality.sites || {};
-export const Filters = state => state.waterquality.filters.characteristics || {};
