--- assets/src/scripts/selectors/wdfn-selector.js
@@ -1,6 +1,0 @@
-import { createSelector } from 'reselect';
-
-export const Sites = state => state.wdfn.sites || {};
-export const siteTypes = state => state.wdfn.filters.siteTypes || {};
-export const Filters = state => state.wdfn.filters || {};
-export const Count = state => state.wdfn.count || {};
