+++ assets/src/scripts/wdfn-home/selectors/wdfn-selector.js
@@ -1,0 +1,6 @@
+import { createSelector } from 'reselect';
+
+export const Sites = state => state.wdfnData.sites || {};
+export const siteTypes = state => state.wdfnData.filters.siteTypes || {};
+export const Filters = state => state.wdfnData.filters || {};
+export const Count = state => state.wdfnData.count || {};
