+++ wdfn-server/waterdata/templates/base_wdfn.html
@@ -1,0 +1,20 @@
+{% extends 'base_plain.html' %}
+
+{% block page_css %}
+      <link rel="stylesheet" href="{{ 'main.css' | asset_url }}">
+      <link rel="stylesheet" href="{{ 'wdfn.css' | asset_url }}">
+{% endblock %}
+
+{% block page_script %}
+    <script src="{{ 'scripts/vendor.js' | asset_url }}"></script>
+    <script async src="{{ 'wdfn-bundle.js' | asset_url }}"></script>
+{% endblock %}
+
+{% block body %}
+    {% include 'partials/base.html' %}
+    <main id="main-content" class="grid-container content-container usa-prose">
+        {% block content %}{% endblock %}
+    </main>
+    {% include 'partials/footer.html' %}
+    {% block extra_body %}{% endblock %}
+{% endblock %}
